backend/requirements.txt
aiofiles==23.1.0
alembic==1.11.1
annotated-types==0.6.0
antlr4-python3-runtime==4.13.1
anyio==3.7.1
asttokens==2.4.1
Authlib==1.2.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
comm==0.2.0
contourpy==1.2.0
cryptography==41.0.4
cycler==0.12.1
decorator==5.1.1
dill==0.3.7
dimod==0.12.12
diskcache==5.6.3
distlib==0.3.7
dnspython==2.4.2
docplex==2.25.236
dwave-cloud-client==0.11.0
dwave-networkx==0.8.14
dwave-preprocessing==0.6.3
dwave-samplers==1.1.0
dwave-system==1.21.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.0
fasteners==0.19
filelock==3.12.4
fonttools==4.45.0
greenlet==3.0.0
h11==0.14.0
homebase==1.0.1
httpcore==0.17.3
httpx==0.24.1
identify==2.5.30
idna==3.4
iniconfig==2.0.0
ipython==8.17.2
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
minorminer==0.2.12
mpmath==1.3.0
networkx==2.8.8
nodeenv==1.8.0
numpy==1.26.1
openqasm3==0.5.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
passlib==1.7.4
pbr==6.0.0
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
plucky==0.4.3
pluggy==1.3.0
ply==3.11
pre-commit==3.3.3
prompt-toolkit==3.0.41
psutil==5.9.6
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pydot==1.4.2
Pygments==2.17.2
pylatexenc==2.10
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.0
python-constraint==1.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
qiskit==0.45.0
qiskit-algorithms==0.2.1
qiskit-optimization==0.6.0
qiskit-qasm3-import==0.4.1
qiskit-terra==0.45.0
rectangle-packer==2.0.2
requests==2.31.0
rsa==4.9
rustworkx==0.13.2
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.13
stack-data==0.6.3
starlette==0.27.0
stevedore==5.1.0
symengine==0.9.2
sympy==1.12
tomli==2.0.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.22.0
virtualenv==20.24.5
wcwidth==0.2.12
Werkzeug==3.0.0
widgetsnbextension==4.0.9
z3-solver==4.12.2.0
docplex
